requirements.txt
flask==3.0.0
flask-cors==4.0.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.24.3
pdfplumber==0.10.3
python-dotenv==1.0.0
requests==2.31.0
gunicorn==21.2.0
